Split RouteState into three partials
This seperates the 3 things that are needed for state.  Kinka gives me Single Responsibility Principle via partials
