+++ source/BlazorState/Features/Routing/RouteState.Behavior.cs
@@ -1,0 +1,18 @@
+﻿namespace BlazorState.Features.Routing
+{
+  using BlazorState;
+
+  public partial class RouteState : State<RouteState>
+  {
+    public RouteState() { }
+
+    protected RouteState(RouteState aState) : this()
+    {
+      Route = aState.Route;
+    }
+
+    public override object Clone() => new RouteState(this);
+
+    protected override void Initialize() { }
+  }
+}
